visual/compare_images.py
"""
Image Comparison Module
Compares screenshots using OpenCV and generates visual diffs.
"""

import cv2
import numpy as np
import imagehash
from PIL import Image

class ImageComparator:
    def __init__(self):
        self.threshold = 0.95
    
    def structural_similarity(self, image1_path, image2_path):
        """Compare images using structural similarity index."""
        pass
    
    def perceptual_hash(self, image1_path, image2_path):
        """Compare images using perceptual hash."""
        pass
    
    def generate_diff_image(self, image1_path, image2_path, output_path):
        """Generate a visual difference image."""
        pass 